Split list item into a mixin
